Two mistakes corrected
I incorrectly used the wrong object name when sending data to the server for notification registration, setting the object name as "channel_id" when it should've been "channel". Additionally, I accidentally believed the current arc domain was "chat-by-arc.github.io" and not "chat-on-arc.github.io". Those changes are reflected here.